+++ app/controllers/api/v2/tasks_controller.rb
@@ -99,6 +99,7 @@ class Api::V2::TasksController < Api::V2Controller
   def permitted_filters
     @permitted_filters ||=
       Task::Filterer::FILTERS_TO_DISPLAY.collect(&:underscore).collect(&:to_sym) +
-      Task::Filterer::FILTERS_TO_HIDE.collect(&:underscore).collect(&:to_sym)
+      Task::Filterer::FILTERS_TO_HIDE.collect(&:underscore).collect(&:to_sym) +
+      [:account_list_id]
   end
 end