+++ app/controllers/concerns/filtering.rb
@@ -1,5 +1,5 @@
 module Filtering
-  CastingError = Class.new(StandardError)
+  DatetimeCastingError = Class.new(StandardError)
   DATE_RANGE_REGEX = /(\d{4}\-\d{2}\-\d{2})(\.\.\.?)(\d{4}\-\d{2}\-\d{2})/
   DATETIME_RANGE_REGEX = /(\d{4}\-\d{2}\-\d{2}T\d{2}\:\d{2}\:\d{2}Z)(\.\.\.?)(\d{4}\-\d{2}\-\d{2}T\d{2}\:\d{2}\:\d{2}Z)/
   DEFAULT_PERMITTED_FILTERS = %w( updated_at ).freeze
@@ -25,8 +25,8 @@ module Filtering
 
   def cast_filter_value(value)
     cast_filter_value!(value)
-  rescue CastingError
-    value
+  rescue DatetimeCastingError
+    raise_bad_request
   end
 
   def cast_filter_value!(value)
@@ -34,8 +34,14 @@ module Filtering
     when DATE_RANGE_REGEX, DATETIME_RANGE_REGEX
       cast_to_datetime_range($LAST_MATCH_INFO)
     else
-      value
+      raise_if_bad_date_range_value(value)
     end
+  end
+
+  def raise_if_bad_date_range_value(value)
+    raise_bad_request if value.to_s.include?('..')
+
+    value
   end
 
   def cast_to_datetime_range(match_data)
@@ -45,6 +51,11 @@ module Filtering
 
     Range.new(start_datetime, end_datetime, exclusive)
   rescue ArgumentError
-    raise CastingError
+    raise DatetimeCastingError
+  end
+
+  def raise_bad_request
+    raise Exceptions::BadRequestError,
+          "Wrong format of date range, should follow 'YYYY-MM-DD...YYYY-MM-DD' for dates and 'YYYY-MM-DDThh:mm:ssZ...YYYY-MM-DDThh:mm:ssZ' for datetimes"
   end
 end